Revert "refactor(stream directories):"
