+++ fanout-fanin/src/main.rs
@@ -19,10 +19,10 @@ fn main() -> Result<(), RuntimeError> {
     let mut rt = Runtime::new().map_err(|_| RuntimeError::Executor)?;
 
     let basis = Instant::now();
-    let timer = source::timer::new(basis.clone(), 500);
-    let (counter_in, counter_out) = transform::counter::new();
-    let (elapsed_time_in, elapsed_time_out) = transform::elapsed_time::new(basis);
-    let stdout = sink::stdout::new();
+    let timer = source::timer(basis.clone(), 500);
+    let (counter_in, counter_out) = transform::counter();
+    let (elapsed_time_in, elapsed_time_out) = transform::elapsed_time(basis);
+    let stdout = sink::stdout();
 
     rt.spawn(cleanup(timer.forward(counter_in.fanout(elapsed_time_in))))
         .spawn(cleanup(
