--- fanout-fanin/src/streams/sink/mod.rs
@@ -1,1 +1,0 @@
-pub mod stdout;
