--- fanout-fanin/src/streams/source/mod.rs
@@ -1,1 +1,0 @@
-pub mod timer;
