+++ fanout-fanin/src/streams/transform.rs
@@ -3,7 +3,7 @@ use std::time::Instant;
 use tokio::prelude::*;
 use utils::stream::connect;
 
-pub fn new() -> (
+pub fn counter() -> (
     impl Sink<SinkItem = Instant, SinkError = RuntimeError>,
     impl Stream<Item = String, Error = RuntimeError>,
 ) {
@@ -19,5 +19,19 @@ pub fn new() -> (
                 format!("count: {}", count)
             }
         }),
+    )
+}
+
+pub fn elapsed_time(
+    basis: Instant,
+) -> (
+    impl Sink<SinkItem = Instant, SinkError = RuntimeError>,
+    impl Stream<Item = String, Error = RuntimeError>,
+) {
+    let (input, output) = connect::<Instant>();
+
+    (
+        input,
+        output.map(move |t: Instant| format!("elapsed time: {:?}", t.duration_since(basis))),
     )
 }
