--- fanout-fanin/src/streams/transform/mod.rs
@@ -1,2 +1,0 @@
-pub mod counter;
-pub mod elapsed_time;
